requirements.txt
flask
flask-sqlalchemy
flask-cors
pyjwt
pydantic
python-dotenv
toml
werkzeug
openai
langchain
langchain-community
langchain-openai
faiss-cpu
pandas
openpyxl
PyPDF2
python-docx
scikit-learn
